move Authenticator to api package + refactor
