+++ styx-api-service/src/main/java/com/spotify/styx/StyxApi.java
@@ -51,7 +51,7 @@ import com.spotify.styx.storage.AggregateStorage;
 import com.spotify.styx.storage.Storage;
 import com.spotify.styx.util.CachedSupplier;
 import com.spotify.styx.util.DockerImageValidator;
-import com.spotify.styx.util.AuthenticatorFactory;
+import com.spotify.styx.api.AuthenticatorFactory;
 import com.spotify.styx.util.StorageFactory;
 import com.spotify.styx.util.Time;
 import com.spotify.styx.util.WorkflowValidator;
