+++ styx-common/src/test/java/com/spotify/styx/util/AuthenticatorFactoryTest.java
@@ -36,7 +36,8 @@ import com.google.api.services.cloudresourcemanager.CloudResourceManager;
 import com.google.api.services.cloudresourcemanager.model.ListProjectsResponse;
 import com.google.api.services.iam.v1.Iam;
 import com.google.common.collect.ImmutableSet;
-import com.spotify.styx.util.AuthenticatorFactory.DefaultAuthenticatorFactory;
+import com.spotify.styx.api.AuthenticatorFactory;
+import com.spotify.styx.api.AuthenticatorFactory.DefaultAuthenticatorFactory;
 import java.io.IOException;
 import org.junit.Before;
 import org.junit.Rule;
